adds a failing test due to xml attribute sequence
demonstrates a simulation of the problem I faced recently
while using XmlApprovals.VerifyXml(...);

I was using XmlApprovals to verify xml which had several namespace
declarations in the root element and these were being serialized
in a different sequence on my build server than on my teams' local
machines.

It appears that the default XmlSerializer class in the .NET
framework uses an unordered collection to store XML attributes
and does not ensure any specific, nor even a consistent,
sequence from one serialization to the next.

I wonder if it might make sense to include an option in the
XmlApprovals.VerifyXml(...) method which allows a consumer
to toggle on some sort of normalizing such that the xml being
verified is normalized in some basic ways (e.g. sequence
of attributes) in order to reduce the chance of false positive
approval failures.