Improve support of xunit2
Default configuration of `FrameworkAssertApproval` does not consider xunit2 at
all. As a result xunit2 tests are failing on continuous integration servers
because ApprovalTests handles them as xunit ones and tries to invoke missing
assert method. The error is the same as in #159.

Besides of making `FrameworkAssertApproval` aware of xunit2 I've added more
smartess to xunit2 reporter itself. Otherwise it conflicted with xunit.